pyproject.toml
[tool.poetry]
name = "tapers-io-flac-tools"
version = "0.1.0"
description = ""
authors = ["0xdead8ead <0xdead8ead@divops.io>"]
readme = "README.md"
packages = [{include = "tapers", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
ffmpeg-python = "^0.2.0"

[tool.poetry.scripts]
flac-tool = "tapers.io.flac.tools.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
